[opt](fe) modify TabletInvertedIndex to local and cloud to reduce memory  (#59683)
